+++ packages/plugins/vite-plugin-generate-shopify-locales/src/utils/getShopLocalization.ts
@@ -1,44 +1,43 @@
 import { createStorefrontClient } from "@solidifront/storefront-client";
 
 type Options = {
-	storeName: string;
-	apiVersion: string;
-	accessToken: string;
+  storeName: string;
+  apiVersion: string;
+  accessToken: string;
 };
 
 const GetShopLocalization = `#graphql
-    query GetShopLocalization {
-      localization {
-        availableCountries {
+  query GetShopLocalization {
+    localization {
+      availableCountries {
+        isoCode
+        name
+        unitSystem
+        availableLanguages {
+          isoCode
+          endonymName
+          name
+        }
+        currency {
           isoCode
           name
-          unitSystem
-          availableLanguages {
-            isoCode
-            endonymName
-            name
-          }
-          currency {
-            isoCode
-            name
-            symbol
-          }
+          symbol
         }
       }
     }
+  }
 ` as const;
 
 export async function getShopLocalization({
-	storeName,
-	apiVersion,
-	accessToken,
+  storeName,
+  apiVersion,
+  accessToken,
 }: Options) {
-	const client = createStorefrontClient({
-		storeName,
-		apiVersion: apiVersion as any,
-		publicAccessToken: accessToken,
-	});
-	const response = await client.query(GetShopLocalization);
-	console.log(response.errors?.graphQLErrors);
-	return response.data!.localization!;
+  const client = createStorefrontClient({
+    storeName,
+    apiVersion: apiVersion as any,
+    publicAccessToken: accessToken,
+  });
+  const response = await client.query(GetShopLocalization);
+  return response.data!.localization!;
 }