+++ packages/start/src/codegen/index.ts
@@ -1,1 +1,0 @@
-export { createSolidifrontConfig } from "@solidifront/codegen";